Split out rate limits
